tests: run slow tests on tagged commits and if @slowtest is present in a commit message
