+++ test-e2e/common/release.test.js
