+++ test-e2e/install/sources.test.js
